# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TranscriptionStatus =
  (typeof TranscriptionStatus)[keyof typeof TranscriptionStatus];

export const TranscriptionStatus = {
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export interface Transcription {
  id: number;
  filename: string;
  fileSize: number;
  text: string;
  status: TranscriptionStatus;
  /** @nullable */
  errorMessage: string | null;
  createdAt: string;
}

export interface TranscriptionStats {
  totalCount: number;
  completedCount: number;
  totalFileSize: number;
}

export type TransformRequestMode =
  (typeof TransformRequestMode)[keyof typeof TransformRequestMode];

export const TransformRequestMode = {
  cleanup: "cleanup",
  rewrite: "rewrite",
} as const;

export type TransformRequestProvider =
  (typeof TransformRequestProvider)[keyof typeof TransformRequestProvider];

export const TransformRequestProvider = {
  deepseek: "deepseek",
  anthropic: "anthropic",
  venice: "venice",
} as const;

export interface TransformRequest {
  mode: TransformRequestMode;
  provider: TransformRequestProvider;
  /** Required when mode is "rewrite". Free-form rewrite directions. */
  instructions?: string;
}

export type TransformResponseMode =
  (typeof TransformResponseMode)[keyof typeof TransformResponseMode];

export const TransformResponseMode = {
  cleanup: "cleanup",
  rewrite: "rewrite",
} as const;

export type TransformResponseProvider =
  (typeof TransformResponseProvider)[keyof typeof TransformResponseProvider];

export const TransformResponseProvider = {
  deepseek: "deepseek",
  anthropic: "anthropic",
  venice: "venice",
} as const;

export interface TransformResponse {
  text: string;
  mode: TransformResponseMode;
  provider: TransformResponseProvider;
}

export interface ErrorResponse {
  error: string;
}

export type CreateTranscriptionBody = {
  /** The audio file to transcribe (MP3, WAV, etc.) */
  file: Blob;
};