# lib/api-zod/src/generated/types/transformRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TransformRequestMode } from "./transformRequestMode";
import type { TransformRequestProvider } from "./transformRequestProvider";

export interface TransformRequest {
  mode: TransformRequestMode;
  provider: TransformRequestProvider;
  /** Required when mode is "rewrite". Free-form rewrite directions. */
  instructions?: string;
}