# lib/api-zod/src/generated/types/transformResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TransformResponseMode } from "./transformResponseMode";
import type { TransformResponseProvider } from "./transformResponseProvider";

export interface TransformResponse {
  text: string;
  mode: TransformResponseMode;
  provider: TransformResponseProvider;
}